Struct rsp2_soa_ops::Part

pub struct Part<L> { /* fields omitted */ }

A partition operator.

It represents a specific way to break up a vector into smaller vectors, each with a label of type L. These labels are usually (but not necessarily) unique, and the partitions are not necessarily contiguous in the original vector. Basically, Part contains a partitioned form of the vector’s indices.

Using the Partition trait, a Part<L> can be applied to any V that implements Partition in order to break it up into a Parted<L, V>.

Implementations

impl<L> Part<L>

pub fn new(part: Parted<L, Vec<usize>>) -> Result<Self, InvalidPartitionError>

Create a partition that decomposes a vector entirely.

Every integer from 0 to the total length must appear once.

pub fn region_keys(&self) -> impl VeclikeIterator<Item = &L>

Iterate over the keys for each region.

Item type is &L;

pub fn region_indices(&self) -> impl VeclikeIterator<Item = &[usize]>

Iterate over the index vectors for each region.

Item type is &[usize];

pub unsafe fn new_unchecked(
part: Parted<L, Vec<usize>>,
index_limit: usize
) -> Self

Safety

Usage of the constructed Part<L> may lead to Undefined Behavior if either of the following conditions are violated:

no usize index may appear more than once in the input.

index_limit must equal the total size.

pub fn from_ord_keys<Ls>(labels: Ls) -> Self where
L: Ord,
Ls: IntoIterator<Item = L>,

Construct a Part from a sequence of keys, with one partition for each distinct value in the sequence.

The partitions in the output will be sorted by key.

pub fn into_parted_indices(self) -> Vec<(L, Vec<usize>)>

pub fn key_vec(&self) -> Vec<&L>

pub fn owned_key_vec(&self) -> Vec<L> where
L: Clone,

pub fn restoring_perm(&self) -> Perm

Get the permutation that would restore the original order of the data, if one were to simply concatenate the partitioned structures produced by applying this Part.