Trait rsp2_soa_ops::Partition

pub trait Partition<'iter>: Sized + 'iter {
    fn into_unlabeled_partitions<L>(
        self, 
        part: &'iter Part<L>
    ) -> Unlabeled<'iter, Self>;

    fn into_partitions<L: Clone>(self, part: &'iter Part<L>) -> Parted<L, Self> { ... }
}

Trait for applying a Part to a Vec (or similar type), breaking it into pieces.

By making this a trait, it can be implemented on types like rsp2’s own Coords or anything else that contains data per-atom (such as eigenvectors).

The lifetime argument ensures that the iterator returned by into_unlabeled_partitions does not outlive either Self or the partition; this allows the iterator to capture self by value, and the partition by reference. Ultimately, we would need both GATs and impl-Trait-on-Trait-impls to get rid of it.

Required methods

fn into_unlabeled_partitions<L>(
self,
part: &'iter Part<L>
) -> Unlabeled<'iter, Self>

Variant of into_partitions which composes more easily, and is therefore the one you need to implement.

It returns an iterator over the partitions of self.

See into_partitions for more info.

Loading content...
Provided methods

fn into_partitions<L: Clone>(self, part: &'iter Part<L>) -> Parted<L, Self>

Consume self to produce partitions.

The ordering within each partition is specified, in order to allow composite types to reliably be able to defer to the implementations defined on each of their members. This is not usually a concern since virtually all implementations will ultimately defer to Vec<_> for their implementation… but in case you must know:

The ordering within each partition of the output must reflect the original order of those elements relative to each other in the input vec, rather than the order of the indices in part.

Loading content...